调整wms中心领域模型

+++ src/main/java/com/zpl/eshop/wms/domain/PurchaseInputOrderItemDTO.java
@@ -40,7 +40,7 @@ public class PurchaseInputOrderItemDTO extends AbstractObject {
     /**
      * 采购价格
      */
-    private Long purchasePrice;
+    private Double purchasePrice;
 
     /**
      * 合格数量
